Implement services folder reorganization according to proposal
MAJOR REORGANIZATION:
- Restructured Services folder from flat 20+ file structure to organized modules
- Created 7 logical service groups: Auth/, Invites/, Navigation/, UI/, Monitoring/, Settings/, Setup/
- Used git mv to preserve file history during reorganization
- Updated all namespace declarations to match new folder structure
- Updated all using statements throughout codebase (20+ files)
- Updated service registrations in Program.cs and ServiceCollectionExtensions.cs

BENEFITS:
- Improved code organization and maintainability
- Clear module boundaries prevent circular dependencies
- Easier navigation and onboarding for developers
- Scalable structure for future service additions
- Related services are co-located for easier updates

STRUCTURE:
- Auth/: User management, roles, admin operations (6 files)
- Invites/: Invitation system services (2 files)
- Navigation/: Menu and navigation services (1 file)
- UI/: Theme and toast notification services (3 files)
- Monitoring/: System health and dashboard services (5 files)
- Settings/: Application settings and encryption (4 files)
- Setup/: First-time setup operations (1 file)

✅ Application builds successfully with no compilation errors
✅ Application starts and runs correctly with all services functional
✅ All git history preserved through git mv operations
