+++ docs/template-enhancement-proposals.md
@@ -1,0 +1,94 @@
+# Template Enhancement Proposals
+
+## High-Impact Infrastructure
+
+### 1. File Management & Media System
+- Blob storage service (local/Azure/AWS)
+- Image processing pipeline
+- Drag-drop upload components
+- Media gallery with metadata
+
+### 2. Real-Time Communication Foundation
+- SignalR hub infrastructure
+- Notification system
+- Chat foundation
+- Event broadcasting
+
+### 3. Flexible Entity Framework Architecture
+- Generic repository pattern
+- Audit trail system
+- Soft delete support
+- Multi-tenant ready
+
+## Business Logic & Workflow
+
+### 4. Scheduling & Calendar System
+- Calendar service with recurring events
+- Reminder system (email/SMS)
+- Public booking interface
+- Time zone handling
+
+### 5. Reporting & Analytics Foundation
+- Chart.js integration
+- Export services (PDF/Excel)
+- Dashboard widgets
+- Usage analytics
+
+### 6. Communication & Integration Layer
+- Email template engine
+- Webhook system
+- External API client base
+- SMS integration
+
+## User Experience & Interface
+
+### 7. Advanced Component Library
+- Data tables with export
+- Dynamic form builder
+- Modal management
+- Search & filter components
+
+### 8. Mobile-First Responsive Framework
+- PWA foundation
+- Touch-friendly controls
+- Responsive navigation
+- Performance monitoring
+
+## Enterprise & Scalability
+
+### 9. Configuration Management System
+- Feature flags
+- Environment-specific config
+- Runtime configuration
+- Configuration validation
+
+### 10. Background Job Processing
+- Hangfire integration
+- Job monitoring
+- Scheduled tasks
+- Queue management
+
+### 11. Advanced Security Framework
+- Content Security Policy
+- Input sanitization
+- API rate limiting
+- Security headers
+
+## Implementation Priority
+
+**Phase 1** (Immediate Impact):
+1. File Management & Media System
+2. Real-Time Communication (SignalR)
+3. Background Job Processing
+4. Advanced Component Library
+
+**Phase 2** (Business Value):
+5. Reporting & Analytics Foundation
+6. Scheduling & Calendar System
+7. Communication & Integration Layer
+8. Mobile-First Responsive Framework
+
+**Phase 3** (Scale & Polish):
+9. Configuration Management System
+10. Flexible Entity Framework Architecture
+11. Advanced Security Framework
