+++ src/blazor-template/Components/Account/Pages/Manage/Index.razor
@@ -5,7 +5,8 @@
 @using BlazorTemplate.Components
 @using BlazorTemplate.Components.Account.Shared
 @using BlazorTemplate.Data
-@using BlazorTemplate.Services
+@using BlazorTemplate.Services.UI
+@using BlazorTemplate.Services.Auth
 @using BlazorTemplate.Components.Shared
 @using Microsoft.AspNetCore.Identity
