+++ src/blazor-template/Components/Account/Pages/Register.razor
@@ -18,7 +18,7 @@
 @inject ILogger<Register> Logger
 @inject NavigationManager NavigationManager
 @inject IdentityRedirectManager RedirectManager
-@inject BlazorTemplate.Services.IInviteService InviteService
+@inject BlazorTemplate.Services.Invites.IInviteService InviteService
 @inject IOptions<BlazorTemplate.Configuration.ConfigurationOptions> ConfigOptions
 
 @layout BlazorTemplate.Components.Account.Shared.AccountLayout
