+++ src/blazor-template/Components/Admin/Dashboard.razor
@@ -1,7 +1,8 @@
 ﻿@page "/admin"
 @using BlazorTemplate.Components.Layout
 @using BlazorTemplate.Data
-@using BlazorTemplate.Services
+@using BlazorTemplate.Services.Monitoring
+@using BlazorTemplate.Services.Auth
 @using BlazorTemplate.Models
 @using BlazorTemplate.Models.Dashboard
 @using Microsoft.AspNetCore.Authorization
