+++ src/blazor-template/Components/Admin/Invites.razor
@@ -2,7 +2,7 @@
 @using Microsoft.AspNetCore.Authorization
 @using Microsoft.AspNetCore.Identity
 @using BlazorTemplate.Data
-@using BlazorTemplate.Services
+@using BlazorTemplate.Services.Invites
 @using BlazorTemplate.Configuration
 @using Microsoft.Extensions.Options
 @using System.ComponentModel.DataAnnotations
