+++ src/blazor-template/Components/Admin/RoleForm.razor
@@ -742,7 +742,7 @@
     </div>
 </div>
 
-@using BlazorTemplate.Services
+@using BlazorTemplate.Services.Auth
 @using BlazorTemplate.Models
 @inject IAdminRoleService AdminRoleService
 @inject IJSRuntime JSRuntime
