+++ src/blazor-template/Components/Admin/RoleUserAssignments.razor
@@ -1,6 +1,6 @@
 @page "/admin/roles/{RoleId}/users"
 @using BlazorTemplate.Components.Layout
-@using BlazorTemplate.Services
+@using BlazorTemplate.Services.Auth
 @using BlazorTemplate.Models
 @using Microsoft.AspNetCore.Authorization
 @layout AdminLayout
