+++ src/blazor-template/Components/Admin/Roles.razor
@@ -671,7 +671,7 @@
     </div>
 </div>
 
-@using BlazorTemplate.Services
+@using BlazorTemplate.Services.Auth
 @using BlazorTemplate.Models
 @inject IAdminRoleService AdminRoleService
 @inject IJSRuntime JSRuntime
