+++ src/blazor-template/Components/Admin/Settings.razor
@@ -1,7 +1,8 @@
 ﻿@page "/admin/settings"
 @using BlazorTemplate.Components.Layout
 @using BlazorTemplate.Data
-@using BlazorTemplate.Services
+@using BlazorTemplate.Services.Settings
+@using BlazorTemplate.Services.UI
 @using BlazorTemplate.Models.Settings
 @using Microsoft.AspNetCore.Authorization
 @using System.Security.Claims
