+++ src/blazor-template/Components/Admin/UserDetails.razor
@@ -2,7 +2,7 @@
 @using BlazorTemplate.Components.Layout
 @using BlazorTemplate.Data
 @using BlazorTemplate.Models
-@using BlazorTemplate.Services
+@using BlazorTemplate.Services.Auth
 @using Microsoft.AspNetCore.Authorization
 @layout AdminLayout
 @attribute [Authorize(Roles = "Administrator")]
