+++ src/blazor-template/Components/App.razor
@@ -1,4 +1,4 @@
-﻿@using BlazorTemplate.Services
+﻿@using BlazorTemplate.Services.UI
 @inject IConfiguration Configuration
 @inject ThemeService ThemeService
