+++ src/blazor-template/Components/Layout/AdminLayout.razor
@@ -1,5 +1,5 @@
 ﻿@inherits LayoutComponentBase
-@using BlazorTemplate.Services
+@using BlazorTemplate.Services.Auth
 @using Microsoft.AspNetCore.Components.Authorization
 @using System.Security.Claims
 @inject AuthenticationStateProvider AuthenticationStateProvider
