+++ src/blazor-template/Components/Layout/MainLayout.razor
@@ -1,4 +1,4 @@
-﻿@using BlazorTemplate.Services
+﻿@using BlazorTemplate.Services.UI
 @inherits LayoutComponentBase
 @implements IDisposable
 @inject IToastService ToastService
