+++ src/blazor-template/Components/Layout/SidebarNavigation.razor
@@ -1,5 +1,6 @@
 ﻿@using BlazorTemplate.Configuration.Navigation
-@using BlazorTemplate.Services
+@using BlazorTemplate.Services.Navigation
+@using BlazorTemplate.Services.Auth
 @using BlazorTemplate.Configuration
 @using Microsoft.AspNetCore.Components.Authorization
 @using System.Security.Claims
