+++ src/blazor-template/Components/Layout/TopNavbar.razor
@@ -1,6 +1,6 @@
 ﻿@using Microsoft.AspNetCore.Components.Authorization
 @using BlazorTemplate.Data
-@using BlazorTemplate.Services
+@using BlazorTemplate.Services.Auth
 @using System.Security.Claims
 @inject AuthenticationStateProvider AuthenticationStateProvider
 @inject IUserRoleService RoleService
