+++ src/blazor-template/Components/Pages/Home.razor
@@ -1,7 +1,7 @@
 ﻿@page "/"
 @rendermode InteractiveServer
 @using BlazorTemplate.Components.Layout
-@using BlazorTemplate.Services
+@using BlazorTemplate.Services.Auth
 @using BlazorTemplate.Models
 @using Microsoft.AspNetCore.Components.Authorization
 @using System.Security.Claims
