+++ src/blazor-template/Components/Shared/ThemeSwitcher.razor
@@ -1,5 +1,5 @@
 @rendermode InteractiveServer
-@using BlazorTemplate.Services
+@using BlazorTemplate.Services.UI
 @inject ThemeService ThemeService
 @inject IJSRuntime JSRuntime
 @implements IDisposable
