+++ src/blazor-template/Program.cs
@@ -3,7 +3,10 @@ using BlazorTemplate.Components.Account;
 using BlazorTemplate.Configuration;
 using BlazorTemplate.Data;
 using BlazorTemplate.Extensions;
-using BlazorTemplate.Services;
+using BlazorTemplate.Services.UI;
+using BlazorTemplate.Services.Settings;
+using BlazorTemplate.Services.Setup;
+using BlazorTemplate.Services.Monitoring;
 using Microsoft.AspNetCore.Components.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
