+++ src/blazor-template/Services/Auth/IUserManagementService.cs
@@ -1,7 +1,7 @@
 using BlazorTemplate.Models;
 using Microsoft.AspNetCore.Identity;
 
-namespace BlazorTemplate.Services
+namespace BlazorTemplate.Services.Auth
 {
     public interface IUserManagementService
     {
